add chip choice
rename to chip choice
