+++ packages/components/src/Chip/ChipChoice.css.d.ts
@@ -1,0 +1,1 @@
+export const wrapper: string;
