+++ packages/components/src/Chip/InternalChip.css
