+++ packages/components/src/Chip/index.ts
@@ -1,3 +1,3 @@
-export { Chip } from "./Chip";
 export { ChipAvatar } from "./ChipAvatar";
 export { ChipIcon } from "./ChipIcon";
+export { ChipChoice, Chip } from "./ChipChoice";
